requirements.txt
asgiref==3.5.2
Django==3.2.13
django-filter==21.1
djangorestframework==3.13.1
importlib-metadata==4.11.3
Markdown==3.3.7
pytz==2022.1
sqlparse==0.4.2
typing_extensions==4.2.0
zipp==3.8.0
